Fixed 2 issues with displaying recent players.
Issue 1: Previous implementation of Comparable broke the way I was
displaying the last 10 Users added. After sorting the ArrayList, the
last 10 Users added are no longer the last 10 elements! A new ArrayList
was created that is never sorted so the last 10 Users added can be
easily accessed.

Issue 2: Every time the "Refresh" button was clicked to update recent
players, new instances of JLabels were being created. This had undesired
outcomes in the GUI. Instantiation of JLabels have been moved outside of
the event listener for the button click so only one instance is created.